throws exception with invalid input
